Lots of misc fixes for building x64 versions of collector and visualizer.

+++ Collector/Allocation.h
@@ -18,7 +18,7 @@ namespace Alloclave
 		};
 
 		void* Address;
-		unsigned int Size;
+		size_t Size;
 		unsigned int Alignment;
 		AllocationType Type;
 		unsigned int HeapId;
