+++ Collector/Registration.h
@@ -7,7 +7,7 @@ namespace Alloclave
 	class CallStack;
 
 	void RegisterTransport(Transport* transport);
-	void RegisterAllocation(void* address, unsigned int size, unsigned int alignment, unsigned int heapId = 0);
+	void RegisterAllocation(void* address, size_t size, unsigned int alignment, unsigned int heapId = 0);
 	void RegisterFree(void* address, unsigned int heapId = 0);
 	void RegisterScreenshot(); // TODO
 	void RegisterCallStackParser(CallStack* parser);
